# be/openapi.yaml
openapi: 3.0.3
info:
  title: cocono-chat backend API
  version: 0.1.0
  description: |
    REST API for cocono-chat (milestones 1–2: accounts + multi-device).

    ## Encoding conventions

    - All binary material (public keys, AES keys, signatures, nonces) is encoded as
      **base64url** (RFC 4648 §5, no padding).
    - Ed25519 public keys are the **raw 32-byte** key (`crypto.subtle.exportKey('raw')`
      in the browser).
    - Signatures are computed over the UTF-8 bytes of the **canonical JSON** of the
      signed object: object keys sorted lexicographically, no whitespace
      (e.g. `canonical({u,p,a,d,t})` → `{"a":...,"d":...,"p":...,"t":...,"u":...}`).
    - Signed payloads (signup, device enrollment) include a client timestamp `t`
      (epoch seconds) that must be within `SIGNED_PAYLOAD_MAX_AGE_SEC` (default 5
      minutes); each signature is additionally de-duplicated in Redis, so captured
      payloads cannot be replayed.

    ## Authentication

    There are no passwords. Login is a challenge-response proof of possession of a
    registered device's Ed25519 private key:

    1. `POST /api/auth/challenge` → one-time nonce `n` (Redis-backed, TTL, single-use).
       A nonce is issued for ANY username/device pair — unknown pairs get a nonce that
       simply never verifies, so the endpoint does not reveal which accounts exist.
    2. Client signs the nonce string (UTF-8 bytes) with the device key.
...
       `Authorization: Bearer <token>`.

...

...

...

...
servers:
...
paths:
...
  /api/devices/enroll:
    post:
      tags: [devices]
      operationId: deviceEnroll
      summary: Request to add this device to an existing account
      description: |
        Same shape and signing rules as signup, but targeting an account that already
        exists and signed by the NEW device's key. The server issues a 6-digit pairing
        code (shown on the new device) that a device already registered on the account
        must approve via `POST /api/devices/approve`. The new device polls
        `GET /api/devices/enroll-status/{enrollId}` until approved.

        Enrollment itself is unauthenticated — approval by an existing device is the
        gate. The payload follows the same freshness + replay rules as signup.
        Rate limited per client IP (default 10/15min).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Enrollment pending approval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollResponse'
        '400':
          description: |
            Validation failure. `error` is one of:
            `invalid_username`, `invalid_device_id`, `invalid_public_key`,
            `invalid_aes_key`, `stale_payload`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: |
            `error` is one of:
            `invalid_signature` (signature does not verify),
            `replay` (this exact signed payload was already used).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'No such account (`error: unknown_account`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            `error` is one of:
            `device_exists` (device id already registered on the account),
            `device_limit` (account already has its maximum number of devices).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/devices/enroll-status/{enrollId}:
    get:
      tags: [devices]
      operationId: deviceEnrollStatus
      summary: Poll the state of a pending enrollment
      description: |
        Called by the enrolling device while it waits for approval. Unauthenticated by
        necessity (the device has no JWT yet); the enrollId is an unguessable
        192-bit capability issued by `POST /api/devices/enroll`.
      parameters:
        - name: enrollId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{16,64}$'
      responses:
        '200':
          description: Enrollment state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollStatusResponse'
        '400':
          description: 'Malformed enrollId (`error: invalid_request`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: 'Enrollment expired or unknown (`error: expired`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/devices/pending:
    post:
      tags: [devices]
      operationId: devicePending
      summary: Inspect a pending enrollment before approving it
      description: |
        Returns the device id and request time for a pairing code, so the approving
        user can see WHAT they are about to approve (defense against social-engineered
        approvals). Does not consume the code.

        Rate limited per account (dedicated bucket, same limits as approve).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
      responses:
        '200':
          description: Pending enrollment details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingResponse'
        '400':
          description: 'Code is not 6 digits (`error: invalid_request`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'Missing, malformed or expired token (`error: unauthorized`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'No pending enrollment with that code for this account (`error: unknown_code`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/devices/approve:
    post:
      tags: [devices]
      operationId: deviceApprove
      summary: Approve a pending enrollment from a registered device
      description: |
        A device already registered on the account enters the 6-digit code shown on the
        new device. Codes are scoped per account and single-use. The new device is
        added atomically, subject to the account's device limit; the enrolling device
        sees the approval on its next status poll.

        Rate limited per account (default 20/15min).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
      responses:
        '200':
          description: Device approved and added to the account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApproveResponse'
        '400':
          description: 'Code is not 6 digits (`error: invalid_request`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'Missing, malformed or expired token (`error: unauthorized`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: |
            No pending enrollment with that code for this account
            (`error: unknown_code`). Also returned when the code was already used or
            expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            `error` is one of:
            `device_exists` (device was added in the meantime),
            `device_limit` (account reached its device maximum).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/devices:
    get:
      tags: [devices]
      operationId: deviceList
      summary: List the authenticated user's devices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Device list (public keys and AES keys are never included).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        '401':
          description: 'Missing, malformed or expired token (`error: unauthorized`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'Account no longer exists (`error: unknown_account`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT issued by /api/auth/verify.
  responses:
    RateLimited:
      description: 'Too many requests (`error: rate_limited`).'
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Username:
      type: string
      description: 5-64 characters of `[a-zA-Z0-9_-]`; normalised to lowercase at signup, case-insensitively unique.
      pattern: '^[a-zA-Z0-9_-]{4,64}$'
      example: alice
    DeviceId:
      type: string
      description: |
        Client-generated device identifier, 8–64 characters of `[a-zA-Z0-9_-]`
        (clients use `crypto.randomUUID()`).
      pattern: '^[a-zA-Z0-9_-]{8,64}$'
      example: 01234567-89ab-cdef-0123-456789abcdef
    PublicKey:
      type: string
      description: Raw 32-byte Ed25519 public key, base64url-encoded.
      example: k7Sb3QKx9d2mVn4rT8uYw1zAe5fGh0jLpQsXcBvN
    AesKey:
      type: string
      description: |
        Raw AES-GCM transport key (16, 24 or 32 bytes; clients use 32),
        base64url-encoded. Encrypts client <-> server traffic only; message content is
        additionally end-to-end encrypted with per-conversation keys.
      example: F9r2KjLm8nPqRsT1uVwXyZa3bCdEfGhIjKlMnOpQrS
    Signature:
      type: string
      description: 64-byte Ed25519 signature, base64url-encoded.
      example: zYxWvUtSrQpOnMlKjIhGfEdCbAzYxWvUtSrQpOnMlKjIhGfEdCbAzYxWvUtSrQpOnMlKjIhGfEdCbA
...
    SignupRequest:
      type: object
      required: [u, p, x, a, d, t, s]
      properties:
        u:
          $ref: '#/components/schemas/Username'
        p:
          $ref: '#/components/schemas/PublicKey'
        a:
          $ref: '#/components/schemas/AesKey'
        d:
          $ref: '#/components/schemas/DeviceId'
        t:
          type: integer
          description: |
            Client timestamp (epoch seconds) at signing time. Must be within
            `SIGNED_PAYLOAD_MAX_AGE_SEC` (default 300s) of server time.
        s:
          allOf:
            - $ref: '#/components/schemas/Signature'
          description: |
            Ed25519 signature over the UTF-8 bytes of the canonical JSON of
            `{ a, d, p, t, u, x }` (keys sorted, no whitespace). Single-use:
            the server rejects any signature it has seen before.
        x:
          allOf:
            - $ref: '#/components/schemas/PublicKey'
          description: |
            Raw 32-byte X25519 key-agreement public key, base64url-encoded
            (milestone 3: pairwise conversation keys). Covered by `s`.
        r:
          allOf:
            - $ref: '#/components/schemas/Username'
          description: |
            OPTIONAL, UNSIGNED referral: the username whose share link
            (`/?chat=<r>`) this signup followed. Recorded as the new account's
            parent (`users.ref`) and as a `created` edge in the admin graph.
            Deliberately outside the signature — widening it would break every
            shipped client for a field that grants nothing (no reward, no
            score, no capability). Junk, self-referrals and unknown names are
            ignored; see docs/SHARES.md.
...
    PairingCode:
      type: string
      description: 6-digit numeric pairing code shown on the enrolling device.
      pattern: '^\d{6}$'
      example: '482913'
    EnrollResponse:
      type: object
      required: [code, enrollId, expiresInSec]
      properties:
        code:
          $ref: '#/components/schemas/PairingCode'
        enrollId:
          type: string
          description: |
            Unguessable capability identifying this enrollment; used to poll
            `GET /api/devices/enroll-status/{enrollId}`.
        expiresInSec:
          type: integer
          description: Seconds until the pairing code expires.
    EnrollStatusResponse:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
          description: |
            True once an existing device approved the enrollment. The enrolling
            device can then log in via the challenge-response flow.
    ApproveRequest:
      type: object
      required: [code]
      properties:
        code:
          $ref: '#/components/schemas/PairingCode'
    ApproveResponse:
      type: object
      required: [approved]
      properties:
        approved:
          allOf:
            - $ref: '#/components/schemas/DeviceId'
          description: The device id that was added to the account.
    PendingResponse:
      type: object
      required: [d, requestedAt]
      properties:
        d:
          $ref: '#/components/schemas/DeviceId'
        requestedAt:
          type: string
          format: date-time
          description: When the enrollment was requested.
    DeviceInfo:
      type: object
      required: [id, main, current, createdAt, lastSeenAt]
      properties:
        id:
          $ref: '#/components/schemas/DeviceId'
        main:
          type: boolean
          description: True for the first (main) device on the account.
        current:
          type: boolean
          description: True when this is the device making the request.
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
    DeviceListResponse:
      type: object
      required: [maxDevices, devices]
      properties:
        maxDevices:
          type: integer
          description: Maximum devices allowed on this account.
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DeviceInfo'
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code (documented per endpoint).
          example: invalid_username
        message:
          type: string
          description: Human-readable description.
          example: Username must be 5+ chars of [a-zA-Z0-9_-]